requirements.txt
streamlit==1.39.0
sqlalchemy==2.0.32
pytesseract==0.3.10
Pillow==10.4.0
chromadb==0.5.11
sentence-transformers==2.7.0
transformers==4.44.2
tokenizers==0.19.1
feedparser==6.0.11
fpdf2==2.7.9
crewai==0.30.11
python-dotenv==1.0.1
pydantic==2.8.2
tabulate==0.9.0
pandas==2.2.2
lxml==5.3.0
pdfplumber
requests
plotly==5.18.0
diskcache==5.6.3
pip freeze
